backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.3.4
SQLAlchemy==2.0.32
alembic==1.13.2
psycopg[binary]==3.2.1
email-validator==2.2.0
httpx==0.27.0
loguru==0.7.2
slowapi==0.1.9
starlette==0.37.2
pytest==8.3.2
pytest-asyncio==0.23.7
faker==26.0.0
freezegun==1.5.1
